requirements.txt
annotated-types==0.7.0
black==24.10.0
click==8.1.8
kafka-python==2.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
py4j==0.10.9.7
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyspark==3.5.4
python-dotenv==1.0.1
typing_extensions==4.12.2
uuid==1.30
